refactor: re-vamp rust core
