+++ phylo2vec/benches/benchmarks/core.rs
@@ -1,6 +1,12 @@
 use criterion::{criterion_group, BenchmarkId, Criterion};
-use phylo2vec::tree_vec::ops;
-use phylo2vec::utils::{sample_matrix, sample_vector};
+use phylo2vec::matrix::base::sample_matrix;
+use phylo2vec::matrix::convert::{
+    from_newick as from_newick_from_matrix, to_newick as to_newick_from_matrix,
+};
+use phylo2vec::vector::base::sample_vector;
+use phylo2vec::vector::convert::{
+    from_newick as from_newick_from_vector, to_newick as to_newick_from_vector,
+};
 use std::ops::RangeInclusive;
 use std::time::Duration;
 
@@ -23,25 +29,25 @@ fn bench_to_newick(c: &mut Criterion) {
 
         // Benchmark ordered case
         group.bench_with_input(
-            BenchmarkId::new("ordered", sample_size),
+            BenchmarkId::new("vector_ordered", sample_size),
             &sample_size,
             |b, &size| {
                 let v = sample_vector(size, true);
 
                 // Benchmark only the to_newick operation
-                b.iter(|| ops::to_newick_from_vector(&v));
+                b.iter(|| to_newick_from_vector(&v));
             },
         );
 
         // Benchmark unordered case
         group.bench_with_input(
-            BenchmarkId::new("unordered", sample_size),
+            BenchmarkId::new("vector_unordered", sample_size),
             &sample_size,
             |b, &size| {
                 let v = sample_vector(size, false);
 
                 // Benchmark only the to_newick operation
-                b.iter(|| ops::to_newick_from_vector(&v));
+                b.iter(|| to_newick_from_vector(&v));
             },
         );
 
@@ -53,7 +59,7 @@ fn bench_to_newick(c: &mut Criterion) {
                 let m = sample_matrix(size, false);
 
                 // Benchmark only the to_newick operation
-                b.iter(|| ops::to_newick_from_matrix(&m.view()));
+                b.iter(|| to_newick_from_matrix(&m.view()));
             },
         );
     }
@@ -61,8 +67,8 @@ fn bench_to_newick(c: &mut Criterion) {
 }
 
 /// Benchmark to_vector
-fn bench_to_vector(c: &mut Criterion) {
-    let mut group = c.benchmark_group("to_vector");
+fn bench_from_newick(c: &mut Criterion) {
+    let mut group = c.benchmark_group("from_newick");
     // Set logarithmic scale for plot
     group.plot_config(
         criterion::PlotConfiguration::default().summary_scale(criterion::AxisScale::Logarithmic),
@@ -70,38 +76,31 @@ fn bench_to_vector(c: &mut Criterion) {
     for i in SAMPLE_SIZES {
         let sample_size = 10000 * i;
         group.bench_with_input(
-            BenchmarkId::from_parameter(sample_size),
+            BenchmarkId::new("vector", sample_size),
             &sample_size,
             |b, &size| {
                 // Generate the Newick string once outside the benchmark loop
                 let v = sample_vector(size, true);
-                let newick = ops::to_newick_from_vector(&v);
+                let newick = to_newick_from_vector(&v);
 
                 // Benchmark only the to_vector operation
-                b.iter(|| ops::to_vector(&newick));
+                b.iter(|| from_newick_from_vector(&newick));
             },
         );
     }
-    group.finish();
-}
 
-fn bench_to_matrix(c: &mut Criterion) {
-    let mut group = c.benchmark_group("to_matrix");
-    // Set logarithmic scale for plot
-    group.plot_config(
-        criterion::PlotConfiguration::default().summary_scale(criterion::AxisScale::Logarithmic),
-    );
     for i in SAMPLE_SIZES {
         let sample_size = 10000 * i;
         group.bench_with_input(
-            BenchmarkId::from_parameter(sample_size),
+            BenchmarkId::new("matrix", sample_size),
             &sample_size,
             |b, &size| {
                 // Generate the matrix once outside the benchmark loop
                 let m = sample_matrix(size, false);
-                let newick = ops::to_newick_from_matrix(&m.view());
+                let newick = to_newick_from_matrix(&m.view());
+
                 // Benchmark only the to_vector operation
-                b.iter(|| ops::matrix::to_matrix(&newick));
+                b.iter(|| from_newick_from_matrix(&newick));
             },
         );
     }
@@ -113,5 +112,5 @@ criterion_group! {
     config = Criterion::default()
         .sample_size(30)
         .warm_up_time(Duration::from_millis(1000));
-    targets = bench_to_newick, bench_to_vector, bench_to_matrix
+    targets = bench_to_newick, bench_from_newick
 }
