+++ phylo2vec/src/lib.rs
@@ -1,2 +1,5 @@
+pub mod matrix;
+pub mod newick;
 pub mod tree_vec;
-pub mod utils;
+mod types;
+pub mod vector;
