+++ phylo2vec/src/profile_main.rs
@@ -1,4 +1,4 @@
-use phylo2vec::tree_vec::ops::{to_newick_from_vector, to_vector};
+use phylo2vec::vector::convert::{from_newick, to_newick};
 use std::env;
 
 const DEFAULT_N_LEAVES: usize = 100000;
@@ -12,7 +12,7 @@ fn main() {
     };
     // For more stability, we use a fixed vector
     let v = (0..(n_leaves - 1)).map(|i| 2 * i).collect::<Vec<usize>>();
-    let n = to_newick_from_vector(&v);
-    let re_v = to_vector(&n);
+    let n = to_newick(&v);
+    let re_v = from_newick(&n);
     println!("vector length: {:?}", re_v.len());
 }
