+++ r-phylo2vec/src/rust/src/lib.rs
@@ -1,10 +1,16 @@
 use extendr_api::prelude::*;
-use ndarray::{Array2, Axis};
+use ndarray::Array2;
 use std::collections::HashMap;
 use std::result::Result;
 
-use phylo2vec::tree_vec::ops;
-use phylo2vec::utils;
+use phylo2vec::matrix::base as mbase;
+use phylo2vec::matrix::convert as mconvert;
+use phylo2vec::matrix::graph as mgraph;
+use phylo2vec::newick;
+use phylo2vec::vector::base as vbase;
+use phylo2vec::vector::convert as vconvert;
+use phylo2vec::vector::graph as vgraph;
+use phylo2vec::vector::ops as vops;
 
 fn as_usize(v: Vec<i32>) -> Vec<usize> {
     v.iter().map(|&x| x as usize).collect()
@@ -53,7 +59,7 @@ fn convert_from_rmatrix(matrix: &Robj) -> Result<Array2<f32>, &'static str> {
 /// @export
 #[extendr]
 fn sample_vector(n_leaves: usize, ordered: bool) -> Vec<i32> {
-    let v = utils::sample_vector(n_leaves, ordered);
+    let v = vbase::sample_vector(n_leaves, ordered);
     as_i32(v)
 }
 
@@ -61,7 +67,7 @@ fn sample_vector(n_leaves: usize, ordered: bool) -> Vec<i32> {
 /// @export
 #[extendr]
 fn sample_matrix(n_leaves: usize, ordered: bool) -> RMatrix<f64> {
-    let matrix = utils::sample_matrix(n_leaves, ordered);
+    let matrix = mbase::sample_matrix(n_leaves, ordered);
     let shape = matrix.shape();
     RMatrix::new_matrix(shape[0], shape[1], |r, c| matrix[[r, c]] as f64)
 }
@@ -71,7 +77,7 @@ fn sample_matrix(n_leaves: usize, ordered: bool) -> RMatrix<f64> {
 #[extendr]
 fn to_newick_from_vector(vector: Vec<i32>) -> String {
     let v_usize = as_usize(vector);
-    ops::to_newick_from_vector(&v_usize)
+    vconvert::to_newick(&v_usize)
 }
 
 /// Recover a rooted tree (in Newick format) from a Phylo2Vec matrix
@@ -79,14 +85,14 @@ fn to_newick_from_vector(vector: Vec<i32>) -> String {
 #[extendr]
 fn to_newick_from_matrix(matrix: RMatrix<f64>) -> String {
     let matrix = convert_from_rmatrix(&matrix).unwrap();
-    ops::to_newick_from_matrix(&matrix.view())
+    mconvert::to_newick(&matrix.view())
 }
 
 /// Convert a newick string to a Phylo2Vec vector
 /// @export
 #[extendr]
 fn to_vector(newick: &str) -> Vec<i32> {
-    let v = ops::to_vector(newick);
+    let v = vconvert::from_newick(newick);
     as_i32(v)
 }
 
@@ -94,7 +100,7 @@ fn to_vector(newick: &str) -> Vec<i32> {
 /// @export
 #[extendr]
 fn to_matrix(newick: &str) -> RMatrix<f64> {
-    let matrix = ops::matrix::to_matrix(newick);
+    let matrix = mconvert::from_newick(newick);
     let shape = matrix.shape();
     RMatrix::new_matrix(shape[0], shape[1], |r, c| matrix[[r, c]] as f64)
 }
@@ -104,7 +110,7 @@ fn to_matrix(newick: &str) -> RMatrix<f64> {
 #[extendr]
 fn check_v(vector: Vec<i32>) {
     let v_usize = as_usize(vector);
-    utils::check_v(&v_usize);
+    vbase::check_v(&v_usize);
 }
 
 /// Validate a Phylo2Vec vector
@@ -113,7 +119,7 @@ fn check_v(vector: Vec<i32>) {
 fn check_m(vector: RMatrix<f64>) {
     let matrix = convert_from_rmatrix(&vector).unwrap();
 
-    utils::check_m(&matrix.view());
+    mbase::check_m(&matrix.view());
 }
 
 /// Get the ancestry matrix of a Phylo2Vec vector
@@ -122,7 +128,7 @@ fn check_m(vector: RMatrix<f64>) {
 fn to_ancestry(vector: Vec<i32>) -> RMatrix<i32> {
     let v_usize: Vec<usize> = as_usize(vector);
     let k = v_usize.len();
-    let ancestry = ops::get_ancestry(&v_usize);
+    let ancestry = vconvert::to_ancestry(&v_usize);
 
     RMatrix::new_matrix(k, 3, |r, c| ancestry[r][c] as i32)
 }
@@ -149,7 +155,7 @@ fn from_ancestry(matrix: RMatrix<i32>) -> Vec<i32> {
         })
         .collect();
 
-    let v = ops::from_ancestry(&ancestry_fixed);
+    let v = vconvert::from_ancestry(&ancestry_fixed);
     as_i32(v)
 }
 
@@ -159,7 +165,7 @@ fn from_ancestry(matrix: RMatrix<i32>) -> Vec<i32> {
 fn to_pairs(vector: Vec<i32>) -> RMatrix<i32> {
     let v_usize: Vec<usize> = as_usize(vector);
     let k = v_usize.len();
-    let pairs = ops::get_pairs(&v_usize);
+    let pairs = vconvert::to_pairs(&v_usize);
     RMatrix::new_matrix(k, 2, |r, c| match c {
         0 => pairs[r].0 as i32,
         1 => pairs[r].1 as i32,
@@ -176,7 +182,7 @@ fn from_pairs(pairs: RMatrix<i32>) -> Vec<i32> {
     let pairs_usize: Vec<(usize, usize)> = (0..nrows)
         .map(|r| (data[r] as usize, data[nrows + r] as usize))
         .collect();
-    let v = ops::from_pairs(&pairs_usize);
+    let v = vconvert::from_pairs(&pairs_usize);
     as_i32(v)
 }
 
@@ -186,7 +192,7 @@ fn from_pairs(pairs: RMatrix<i32>) -> Vec<i32> {
 fn to_edges(vector: Vec<i32>) -> RMatrix<i32> {
     let v_usize: Vec<usize> = as_usize(vector);
     let k = v_usize.len();
-    let edges = ops::get_edges(&v_usize);
+    let edges = vconvert::to_edges(&v_usize);
     RMatrix::new_matrix(2 * k, 2, |r, c| match c {
         0 => edges[r].0 as i32,
         1 => edges[r].1 as i32,
@@ -203,7 +209,7 @@ fn from_edges(edges: RMatrix<i32>) -> Vec<i32> {
     let edges_usize: Vec<(usize, usize)> = (0..nrows)
         .map(|r| (data[r] as usize, data[nrows + r] as usize))
         .collect();
-    let v = ops::from_edges(&edges_usize);
+    let v = vconvert::from_edges(&edges_usize);
     as_i32(v)
 }
 
@@ -212,7 +218,7 @@ fn from_edges(edges: RMatrix<i32>) -> Vec<i32> {
 #[extendr]
 fn add_leaf(vector: Vec<i32>, leaf: i32, branch: i32) -> Vec<i32> {
     let mut v_usize: Vec<usize> = as_usize(vector);
-    let new_v = ops::add_leaf(&mut v_usize, leaf as usize, branch as usize);
+    let new_v = vops::add_leaf(&mut v_usize, leaf as usize, branch as usize);
     as_i32(new_v)
 }
 
@@ -221,7 +227,7 @@ fn add_leaf(vector: Vec<i32>, leaf: i32, branch: i32) -> Vec<i32> {
 #[extendr]
 fn remove_leaf(vector: Vec<i32>, leaf: i32) -> Robj {
     let mut v_usize: Vec<usize> = as_usize(vector);
-    let (new_v, branch) = ops::remove_leaf(&mut v_usize, leaf as usize);
+    let (new_v, branch) = vops::remove_leaf(&mut v_usize, leaf as usize);
     list!(v = as_i32(new_v), branch = branch as i32).into()
 }
 
@@ -231,9 +237,9 @@ fn remove_leaf(vector: Vec<i32>, leaf: i32) -> Robj {
 fn cophenetic_from_vector(vector: Vec<i32>) -> RMatrix<i32> {
     let v_usize: Vec<usize> = as_usize(vector);
     let k = v_usize.len();
-    let coph_usize = ops::cophenetic_distances(&v_usize);
+    let coph_usize = vgraph::cophenetic_distances(&v_usize);
     let mut coph = RMatrix::new_matrix(k + 1, k + 1, |r, c| coph_usize[[r, c]] as i32);
-    let dimnames = (0..k + 1).map(|x| x as i32).collect::<Vec<i32>>();
+    let dimnames = (0..=k).map(|x| x as i32).collect::<Vec<i32>>();
     coph.set_dimnames(List::from_values(vec![dimnames.clone(), dimnames]));
 
     coph
@@ -245,7 +251,7 @@ fn cophenetic_from_vector(vector: Vec<i32>) -> RMatrix<i32> {
 fn cophenetic_from_matrix(matrix: RMatrix<f64>) -> RMatrix<f64> {
     let matrix_f32 = convert_from_rmatrix(&matrix).unwrap();
     let k = matrix_f32.shape()[0];
-    let coph_f32 = ops::matrix::cophenetic_distances_with_bls(&matrix_f32.view());
+    let coph_f32 = mgraph::cophenetic_distances(&matrix_f32.view());
     let mut coph = RMatrix::new_matrix(k + 1, k + 1, |r, c| coph_f32[[r, c]] as f64);
     let dimnames = (0..k + 1).map(|x| x as i32).collect::<Vec<i32>>();
     coph.set_dimnames(List::from_values(vec![dimnames.clone(), dimnames]));
@@ -257,7 +263,7 @@ fn cophenetic_from_matrix(matrix: RMatrix<f64>) -> RMatrix<f64> {
 /// @export
 #[extendr]
 fn has_branch_lengths(newick: &str) -> bool {
-    ops::newick::has_branch_lengths(newick)
+    newick::has_branch_lengths(newick)
 }
 
 /// Create an integer-taxon label mapping (label_mapping)
@@ -271,7 +277,7 @@ fn has_branch_lengths(newick: &str) -> bool {
 /// @export
 #[extendr]
 fn create_label_mapping(newick: &str) -> List {
-    let (nw_int, label_mapping) = ops::newick::create_label_mapping(newick);
+    let (nw_int, label_mapping) = newick::create_label_mapping(newick);
 
     let mut label_mapping_list: Vec<String> = Vec::new();
 
@@ -302,7 +308,7 @@ fn apply_label_mapping(newick: &str, label_mapping_list: Vec<String>) -> String 
         .collect::<HashMap<usize, String>>();
 
     // Not great, but cannot figure out how to map_err to an rextendr error
-    ops::newick::apply_label_mapping(newick, &label_mapping_hash).unwrap()
+    newick::apply_label_mapping(newick, &label_mapping_hash).unwrap()
 }
 
 // Get the first recent common ancestor between two nodes in a Phylo2Vec tree
@@ -314,8 +320,7 @@ fn apply_label_mapping(newick: &str, label_mapping_list: Vec<String>) -> String 
 #[extendr]
 fn get_common_ancestor(vector: Vec<i32>, node1: i32, node2: i32) -> i32 {
     let v_usize: Vec<usize> = as_usize(vector);
-    let common_ancestor =
-        ops::vector::get_common_ancestor(&v_usize, node1 as usize, node2 as usize);
+    let common_ancestor = vops::get_common_ancestor(&v_usize, node1 as usize, node2 as usize);
     common_ancestor as i32
 }
 
@@ -336,7 +341,7 @@ fn get_common_ancestor(vector: Vec<i32>, node1: i32, node2: i32) -> i32 {
 #[extendr]
 fn queue_shuffle(vector: Vec<i32>, shuffle_cherries: bool) -> List {
     let v_usize: Vec<usize> = as_usize(vector);
-    let (v_new, label_mapping) = ops::vector::queue_shuffle(&v_usize, shuffle_cherries);
+    let (v_new, label_mapping) = vops::queue_shuffle(&v_usize, shuffle_cherries);
     list!(v = as_i32(v_new), mapping = as_i32(label_mapping))
 }
